GRIND-50: Only delete the instance if --leave_on_failure is false

+++ grinder/harness.py
@@ -393,7 +393,8 @@ class TestHarness(Notifier):
             if host is not None:
                 assert host.id == instance.get_host().id
         except:
-            instance.delete()
+            if not(self.config.leave_on_failure):
+                instance.delete()
             raise
         if agent:
             try:
